i updated the route  name so as to fit in to the right pae
