+++ styles/globals.css
@@ -63,9 +63,9 @@
 
 
 .ql-editor {
-  /* text-align: center; 
+  text-align: center; 
   background-color: white;
-  padding: 10px; */
+  padding: 10px;
 }
 
 @layer utilities {
